+++ packages/desktop/Cargo.toml
@@ -50,7 +50,7 @@ tao = { workspace = true, features = ["rwh_05"] }
 once_cell = { workspace = true }
 dioxus-history = { workspace = true }
 base64 = { workspace = true }
-
+libc = "0.2.170"
 
 [target.'cfg(unix)'.dependencies]
 signal-hook = "0.3.17"
