+++ packages/dioxus/src/launch.rs
@@ -332,12 +332,12 @@ impl LaunchBuilder {
 
         #[cfg(feature = "mobile")]
         if matches!(platform, KnownPlatform::Mobile) {
-            return dioxus_mobile::launch_bindings::launch(app, contexts, configs);
+            return dioxus_desktop::launch::launch_cfg(app, contexts, configs);
         }
 
         #[cfg(feature = "desktop")]
         if matches!(platform, KnownPlatform::Desktop) {
-            return dioxus_desktop::launch::launch(app, contexts, configs);
+            return dioxus_desktop::launch::launch_cfg(app, contexts, configs);
         }
 
         #[cfg(feature = "server")]
